+++ package.json
@@ -7,7 +7,8 @@
     "dev": "webpack server --config webpack/dev.js",
     "start": "npm run build && node dist/server/index.js",
     "build:client": "webpack --config webpack/prod.js",
-    "build:server": "tsc --project server",
+    "build:server": "tsc --project ./src/server && tsc-alias",
+    "start:server": "node dist/server/index.js",
     "build": "npm run build:client && npm run build:server",
     "build:analyze": "webpack --config webpack/prod.js --env ANALYZE",
     "lint": "eslint src",
@@ -38,7 +39,8 @@
     "react-dom": "^17.0.2",
     "react-hot-loader": "^4.13.0",
     "react-redux": "^8.0.2",
-    "react-router-dom": "^6.3.0"
+    "react-router-dom": "^6.3.0",
+    "styled-components": "^5.3.5"
   },
   "engines": {
     "node": "16.15.0"
@@ -50,6 +52,7 @@
     "@types/react": "^17.0.45",
     "@types/react-dom": "^17.0.17",
     "@types/serviceworker": "^0.0.47",
+    "@types/styled-components": "^5.1.25",
     "@typescript-eslint/eslint-plugin": "^5.27.0",
     "@typescript-eslint/parser": "^5.26.0",
     "circular-dependency-plugin": "^5.2.2",
@@ -81,6 +84,7 @@
     "style-loader": "^3.3.1",
     "terser-webpack-plugin": "^5.3.1",
     "ts-loader": "^9.3.0",
+    "tsc-alias": "^1.6.11",
     "tsc-files": "^1.1.3",
     "typescript": "^4.7.2",
     "webpack": "^5.72.1",
