+++ src/layout/public-layout/public-layout.tsx
@@ -1,4 +1,4 @@
 import type { FC } from 'react';
-import './public-layout.css';
+// import './public-layout.css';
 
 export const PublicLayout: FC = ({ children }) => <div>{children}</div>;
