+++ src/pages/game/components/bird/bird.ts
@@ -1,5 +1,5 @@
+import bird from 'assets/images/bird.png';
 import { PipeConstants } from 'pages/game/components/pipe';
-import bird from '../../../../assets/images/bird.png';
 import type { PipesPositionsConfig } from '../pipe/pipe';
 import { BirdConstants } from './bird-constants';
