+++ src/pages/game/game.tsx
@@ -1,5 +1,5 @@
 import { GameBoard } from './components/game-board';
-import './game.css';
+// import './game.css';
 
 export const Game = () => (
   <div>
