+++ src/pages/plug/plug.tsx
@@ -1,5 +1,5 @@
 import { CircularProgress } from '@mui/material';
-import './plug.css';
+// import './plug.css';
 
 export const PlugComponent = () => (
   <div className="loader-container">
