+++ src/pages/signin/signin-page.tsx
@@ -1,6 +1,6 @@
 import { CardFormComponent } from 'components/card-form';
 import { SigninForm } from './components/signin-form';
-import './signin-page.css';
+// import './signin-page.css';
 
 export const SignInPage = () => (
   <div className="card-form-wrapper">
