+++ src/pages/signup/signup-page.tsx
@@ -1,6 +1,6 @@
 import { CardFormComponent } from 'components/card-form';
 import { SignupForm } from './components/signup-form';
-import './signup-page.css';
+// import './signup-page.css';
 
 export const SignUpPage = () => (
   <div className="card-form-wrapper">
