+++ src/pages/welcome-page/welcome-page.tsx
@@ -1,5 +1,5 @@
 import { InfoAboutProject } from './components/info-about-project';
-import './welcome-page.css';
+// import './welcome-page.css';
 
 export const WelcomePage = () => (
   <div className="main-page">
