+++ src/server/render/hmr.ts
@@ -1,0 +1,22 @@
+// import webpack from 'webpack';
+// import config from '../../../webpack/common';
+// import devMiddleware from 'webpack-dev-middleware';
+// // @ts-ignore
+// import hotMiddleware from '@gatsbyjs/webpack-hot-middleware'
+// import { render } from './render';
+
+// const compiler = webpack(config);
+
+// export default [
+//   devMiddleware(compiler, {
+//     serverSideRender: true,
+//     index: false,
+//     publicPath: config.output!.publicPath!,
+//   }),
+//   hotMiddleware(compiler, {
+//     path: `/__webpack_hmr`,
+//     log: false,
+//     heartbeat: 10 * 1000,
+//   }),
+//   render
+// ]
