+++ src/server/render/index.ts
@@ -1,0 +1,7 @@
+// export default () => {
+//   if (process.env.NODE_ENV === 'development') {
+//     return require('./hmr').default
+//   } else {
+//     return require('./render').default;
+//   }
+// }
